+++ bettersocket/__init__.py
@@ -2,6 +2,6 @@
 
 __author__ = """Marcsello Hooves"""
 __email__ = 'punkosdmarcell@rocketmail.com'
-__version__ = '0.1.0'
+__version__ = '0.1.1'
 
 from .bettersocket import BetterSocketIO, BetterSocketReader, BetterSocketWriter