+++ bettersocket/bettersocket.py
@@ -4,15 +4,17 @@ import select
 from typing import Optional
 
 
-class BetterSocketReader():
+class BetterSocketReader(object):
     """
-    This socket reader is designed to read delimited chunks from the socket efficently
+    This is a wrapper for low-level sockets, for reading delimited frames.
+
+    Both blocking and non-blocking sockets work out of the box.
     """
 
     def __init__(self, sock: socket.socket, delimiter: bytes = b"\n"):
 
         if len(delimiter) != 1:
-            raise ProgrammingError("Delimiter must be 1 byte long")
+            raise ValueError("Delimiter must be 1 byte long")
 
         if not isinstance(sock, socket.socket):
             raise TypeError("Socket must be an instance of socket.socket")
@@ -35,25 +37,25 @@ class BetterSocketReader():
 
     def reset(self):
         """
-        This call clears the internal buffer of the BetterSocketReader instance.
-        It does not alter the kernel buffer.
+        This call clears the internal buffer of the instance.
+        This does not clear the kernel buffer.
         """
         self._buffer = bytes()
 
-    def readframe(self, chunksize: int = 1024) -> Optional[str]:  # timeouting and nonblocking sockets both expected as well as blocking, expected to be called in a loop
+    def readframe(self, chunksize: int = 1024) -> Optional[str]:
         """
-        Returns one chunk of data between delimiters (without the delimiters)
-        Returns None if nothing to read (no delimiter recieved)
+        Returns one frame of data between delimiters (without the delimiters)
+        Returns None if nothing to read (no delimiter received)
         """
 
-        data = self._pop_one_from_buffer()  # before recieve, check if there is a valid data in the buffer
+        data = self._pop_one_from_buffer()  # before receive, check if there is a valid data in the buffer
         if data:
             return data
 
-        # actual recieving won't start until there is no more valid message left in the buffer
+        # actual receiving won't start until there is no more valid message left in the buffer
 
         try:
-            chunk = self._sock.recv(chunksize)  # recieve a chunk
+            chunk = self._sock.recv(chunksize)  # receive a chunk
         except socket.timeout:
             return None
         except socket.error as e:
@@ -63,16 +65,17 @@ class BetterSocketReader():
                 raise  # everything else should be raised
 
         if chunk:
-            self._buffer += chunk  # append the recieved chunk to the buffer
-            return self._pop_one_from_buffer()  # and check if a valid message recieved
+            self._buffer += chunk  # append the received chunk to the buffer
+            return self._pop_one_from_buffer()  # and check if a valid message received
         else:
             raise ConnectionResetError()  # chunk is only none when the connection is dropped (otherwise it would have returned)
 
 
-class BetterSocketWriter():
+class BetterSocketWriter(object):
     """
-    This is a wrapper for sending with nonblocking sockets
-    It kinda makes a nonblocking socket a little blocking in terms of sending, avoiding socket not ready errors
+    This is a wrapper for low-level sockets, for sending delimited frames.
+
+    Both blocking and non-blocking sockets supported out of the box. Either way it waits for the socket to became ready.
     """
 
     def __init__(self, sock: socket.socket, delimiter: bytes = b"\n"):
@@ -85,7 +88,8 @@ class BetterSocketWriter():
 
     def rawsendall(self, data: bytes):
         """
-        This call is blocks until the socket is ready to send data. Then sends the data.
+        This call is blocks until the socket is ready. Then sends the data.
+        Does not append the delimiter.
         """
 
         writable = select.select([], [self._sock], [])[1]
@@ -95,17 +99,17 @@ class BetterSocketWriter():
 
     def sendframe(self, data: bytes):
         """
-        This call sends a frame.
-        Works the same as rawsendall
+        This call automatically appends the delimiter to the end of the data.
+        blocks until the socket is ready.
         """
         self.rawsendall(data + self._delimiter)
 
 
-class BetterSocketIO:
+class BetterSocketIO(object):
     """
-    This class combines BetterSocketReader and BetterSocketWriter together
-    This is a simple wrapper that makes working with sockets just a little bit easier
-    Also exposes the fileno (as well as some basic functionality) so this class can be used in select.select
+    This class combines BetterSocketReader and BetterSocketWriter together.
+    Functions from both classes are exposed.
+    This is the recommended wrapper to use.
     """
 
     def __init__(self, sock: socket.socket, delimiter: bytes = b"\n"):
@@ -115,9 +119,15 @@ class BetterSocketIO:
         self._writer = BetterSocketWriter(sock, delimiter)
 
     def readframe(self, chunksize: int = 1024) -> Optional[bytes]:
+        """
+        Same as BetterSocketReader.readframe
+        """
         return self._reader.readframe(chunksize)
 
     def rawsendall(self, data: bytes):
+        """
+        Same as BetterSocketWriter.rawsendall
+        """
         self._writer.rawsendall(data)
 
     def sendframe(self, data: bytes):
@@ -126,8 +136,11 @@ class BetterSocketIO:
         """
         self._writer.sendframe(data)
 
-    def fileno(self) -> int:
-        return self._socket.fileno()
+    def reset(self):
+        """
+        Same as BetterSocketReader.reset
+        """
+        self._reader.reset()
 
     def close(self):
         """
@@ -135,13 +148,21 @@ class BetterSocketIO:
         After this call no further calls should be attempted.
         """
         self._socket.close()
-        self._reader = None  # are those really necessary?
+        self._reader = None
         self._writer = None
 
-    def __str__(self):  # This is probably not very good for UINIX sockets
+    def __str__(self) -> str:
         try:
-            host, port = self._socket.getpeername()
-            return "Socket connected to {}:{}".format(host, port)
+
+            if self._socket.family in [socket.AF_INET, socket.AF_INET6]:
+                host, port = self._socket.getpeername()
+                return f"Socket connected to {host}:{port}"
+
+            else:  # including AF_UNIX
+                return f"Socket connected to {self._socket.getpeername()}"
 
         except (OSError, BrokenPipeError):
             return "Unconnected socket"
+
+    def __repr__(self) -> str:
+        return f"<{str(self)}>"